fix: long indexing migration batch 4 - MEDIUM priority issues
M1: IL kernel comments updated from int* to long* (5 files)
- ILKernelGenerator.Comparison.cs (2 locations)
- ILKernelGenerator.MixedType.cs
- ILKernelGenerator.Reduction.cs
- ILKernelGenerator.Unary.cs

M4: Confirmed complete - long* version exists in Shape.Unmanaged.cs
- Added cross-reference comment to int* version in Shape.cs

M6: np.load int total accumulator changed to long
M7: np.save int total accumulator changed to long
M9: Confirmed already fixed - NDArray<T> has long size constructors

New issues found and fixed:
- M12: np.random.randn loop counter int -> long
- M13: ILKernelGenerator.Scan.cs loop counters int -> long
  (42 replacements: outer, inner, i over outerSize/innerSize/axisSize)

Updated LONG_INDEXING_ISSUES.md with batch 4 status

+++ src/NumSharp.Core/APIs/np.save.cs
@@ -73,7 +73,7 @@ namespace NumSharp
 
         private static ulong writeValueMatrix(BinaryWriter reader, Array matrix, int bytes, int[] shape)
         {
-            int total = 1;
+            long total = 1;
             for (int i = 0; i < shape.Length; i++)
                 total *= shape[i];
             var buffer = new byte[bytes * total];
